Replace hint check to text view
