+++ kakao/src/main/kotlin/com/agoda/kakao/text/TextViewAssertions.kt
@@ -118,4 +118,24 @@ interface TextViewAssertions : BaseAssertions {
         view.check(ViewAssertions.matches(
                 ViewMatchers.withText(Matchers.startsWith(text))))
     }
+
+    /**
+     * Checks if the view has given hint
+     *
+     * @param hint Text to be matched
+     */
+    fun hasHint(hint: String) {
+        view.check(ViewAssertions.matches(
+                ViewMatchers.withHint(hint)))
+    }
+
+    /**
+     * Checks if the view has given hint
+     *
+     * @param resId String resource to be matched
+     */
+    fun hasHint(@StringRes resId: Int) {
+        view.check(ViewAssertions.matches(
+                ViewMatchers.withHint(resId)))
+    }
 }
